refactor: extract SettingsGlobalHotkeysPane.swift
Extracts the SettingsGlobalHotkeysPane View (57 lines) from
SettingsGeneralPanes.swift into its own file.

Byte-identical move of lines 52-108. Only imports SwiftUI.

Reduces SettingsGeneralPanes.swift from 108 to 50 lines. Tests: 667.

⚠️ Manual smoke check recommended: render Settings > Global Hotkeys pane.
